Update README.md
@RobertLRead I am working to update the readme per the marketing plan document.
I rearranged some of the sections, added that the GPAD is currently for sale, and a link to the published article from HardwareX. I also referenced the Krake as the wireless version of the project. I can also move some of the photos from the article to the readme. 
Also, I noticed that the Summer 2023 updates seem to show a new version of the GPAD-could I label it as Version 3? I'm not sure if there were any recent changes or progress after 2023, which I could add to the end of the readme. Finally, I'm not sure if the volunteer section is up to date.
Thank you,
Mairin O'Grady